chore(build): trigger downstream repo release
OpenEBS Control Plane code is split across multiple repos.

For releasing control plane code, multiple repo's need
to be tagged in a certain order.

This fix will help to trigger the release
on the downstream repo. In this case, after
releasing openebs/maya, the following repos should be
tagged:
  - openebs/velero-plugin
  - openebs/cstor-csi
  - openebs/jiva-operator

Note: Due to multiple jobs running in travis,
the release on the downstream should be done
only after the longest job that pushes the
containers is completed. This is manually
configured via ENV variable RELEASE_DOWNSTREAM.

Signed-off-by: kmova <kiran.mova@mayadata.io>