fix(maya-exporter): fix duplicate metrics for openebs_zpool_last_sync_time (#1678)
This PR makes the labeled pool value as a pod name, which means the label will be the same in case of error and non-error. So that it does not lead to two different metrics. It also does not set the last_sync_time value in case of error so as to retain the previous metric value.

Signed-off-by: Sumit Lalwani <sumit.lalwani97@gmail.com>