fix(usage): add nil checks to avoid panic
Once a PVC create request is processed by the CAST,
the response is inspected to fetch the PVC name
and other details to generate an analytics log.

This was done as part of the PR: https://github.com/openebs/maya/pull/1708

This commit adds a nil check to handle the case of error response
to avoid a panic.

Signed-off-by: kmova <kiran.mova@mayadata.io>